refactor(frontend): extract customer api module
